+++ backend/src/nodes/ncnn_nodes.py
@@ -17,7 +17,7 @@ from .node_factory import NodeFactory
 from .properties.inputs import *
 from .properties.outputs import *
 from .utils.ncnn_auto_split import ncnn_auto_split_process
-from .utils.ncnn_structure import NcnnModel
+from .utils.ncnn_model import NcnnModel
 from .utils.utils import get_h_w_c, convenient_upscale
